mail_service/requirements.txt
fastapi
uvicorn
aiosmtplib
python-dotenv
email-validator
